+++ modules/alloy/changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fixed `TouchDragging` behaviour triggering drag on any document touchmove event.
+- Fixed dragging updating start state on window scroll, when dragging hadn't started.
 
 # [5.0.1] - 2019-10-25